Switched doubles to floats, to solve differences with Java version in Pittsburgh crossover

+++ src/mofgbmlpy/data/input.py
@@ -44,7 +44,7 @@ class Input:
             pattern_id = 0
             patterns = []
             for row in reader:
-                in_vector = np.zeros(num_dim, dtype=np.float64)
+                in_vector = np.zeros(num_dim, dtype=np.float32)
                 c_vector = np.zeros(num_classes, dtype=np.int_)
 
                 for i in range(len(in_vector)):
@@ -86,7 +86,7 @@ class Input:
             patterns = []
 
             for row in reader:
-                in_vector = np.zeros(num_dim, dtype=np.float64)
+                in_vector = np.zeros(num_dim, dtype=np.float32)
 
                 for i in range(len(in_vector)):
                     in_vector[i] = float(row[i])
