+++ src/mofgbmlpy/fuzzy/fuzzy_term/fuzzy_set/fuzzy_set.pxd
@@ -7,11 +7,11 @@ cdef class FuzzySet:
     cdef int __id
     cdef int __division_type
 
-    cdef double get_membership_value(self, double x)
+    cdef float get_membership_value(self, float x)
     cpdef get_term(self)
     cpdef get_function_callable(self)
     cpdef int get_id(self)
     cpdef AbstractMF get_function(self)
     cpdef set_function(self, AbstractMF function)
     cpdef get_division_type(self)
-    cpdef double get_support(self, double x_min=?, double x_max=?)
+    cpdef float get_support(self, float x_min=?, float x_max=?)
