+++ src/mofgbmlpy/fuzzy/fuzzy_term/fuzzy_variable.pxd
@@ -6,13 +6,13 @@ import cython
 cdef class FuzzyVariable:
     cdef FuzzySet[:] __fuzzy_sets
     cdef str __name
-    cdef double[:] __domain
+    cdef float[:] __domain
 
     cpdef str get_name(self)
-    cdef double get_membership_value(self, int fuzzy_set_index, double x)
+    cdef float get_membership_value(self, int fuzzy_set_index, float x)
     cpdef int get_length(self)
     cpdef FuzzySet get_fuzzy_set(self, int fuzzy_set_index)
-    cpdef double get_support(self, int fuzzy_set_index)
+    cpdef float get_support(self, int fuzzy_set_index)
     cpdef get_fuzzy_sets(self)
     cpdef get_support_values(self)
     cpdef get_domain(self)
