+++ src/mofgbmlpy/fuzzy/fuzzy_term/membership_function/abstract_mf.pyx
@@ -5,34 +5,34 @@ from mofgbmlpy.exception.abstract_method_exception import AbstractMethodExceptio
 
 cdef class AbstractMF:
     """Abstract membership function
-        _params (double[]): List of parameters values
+        _params (float[]): List of parameters values
         _are_params_points_flag (bool): If true then we can move them in an interactive plot. e.g. for gaussian it's set to false
     """
-    def __init__(self, double[:] params, bint are_params_points_flag=True):
+    def __init__(self, float[:] params, bint are_params_points_flag=True):
         self._params = params
         if self._params is None:
-            self._params = np.empty(0)
+            self._params = np.empty(0, dtype=np.float32)
         self._are_params_points_flag = are_params_points_flag
 
-    cdef double get_value(self, double x):
+    cdef float get_value(self, float x):
         """Get membership value (accessible only from Cython code)
         
         Args:
-            x (double): Value whose membership value is calculated
+            x (float): Value whose membership value is calculated
 
         Returns:
-            double: Membership value
+            float: Membership value
         """
         raise AbstractMethodException()
 
-    def get_value_py(self, double x):
+    def get_value_py(self, float x):
         """Get membership value
 
         Args:
-            x (double): Value whose membership value is calculated
+            x (float): Value whose membership value is calculated
 
         Returns:
-            double: Membership value
+            float: Membership value
         """
 
         return self.get_value(x)
@@ -63,40 +63,40 @@ cdef class AbstractMF:
 
         return root
 
-    cpdef cnp.ndarray[double, ndim=1] get_params(self):
+    cpdef cnp.ndarray[float, ndim=1] get_params(self):
         """Get the parameters of this function (as a numpy array)
         
         Returns:
-            double[]: Parameters
+            float[]: Parameters
         """
-        return np.array(self._params)
+        return np.array(self._params, dtype=np.float32)
 
-    cpdef cnp.ndarray[double, ndim=1] get_param_range(self, int index, double x_min=0, double x_max=1):
+    cpdef cnp.ndarray[float, ndim=1] get_param_range(self, int index, float x_min=0, float x_max=1):
         """Get the range of acceptable values a given parameter as a numpy array of two values
         
         Args:
             index (int): Index of the parameter whose range is got
-            x_min (double): Min value of the domain for the x axis (e.g. if index is 0 for the triangular set then we get [xmin, center]
-            x_max (double): Max value of the domain for the x axis (e.g. if index is 2 for the triangular set then we get [center, max]
+            x_min (float): Min value of the domain for the x axis (e.g. if index is 0 for the triangular set then we get [xmin, center]
+            x_max (float): Max value of the domain for the x axis (e.g. if index is 2 for the triangular set then we get [center, max]
 
         Returns:
-            double[]: Range of possible values
+            float[]: Range of possible values
         """
         raise AbstractMethodException()
 
-    cpdef bint is_param_value_valid(self, int index, double value, double x_min=0, double x_max=1):
+    cpdef bint is_param_value_valid(self, int index, float value, float x_min=0, float x_max=1):
         """Check if the provided value for the parameter at the given index is valid
         
         Args:
             index (int): Index of the parameter
-            value (double): Value that is checked for the parameter
-            x_min (double): Min value of the domain for the x axis
-            x_max (double): Max value of the domain for the x axis
+            value (float): Value that is checked for the parameter
+            x_min (float): Min value of the domain for the x axis
+            x_max (float): Max value of the domain for the x axis
 
         Returns:
             bool: True if it is valid and false otherwise
         """
-        cdef double[:] val_range = self.get_param_range(index, x_min, x_max)
+        cdef float[:] val_range = self.get_param_range(index, x_min, x_max)
         return val_range[0] <= value and value <= val_range[1]
     
     cpdef bint are_params_points(self):
@@ -107,14 +107,14 @@ cdef class AbstractMF:
         """
         return self._are_params_points_flag
 
-    cpdef void set_param_value(self, int index, double value, double x_min=0, double x_max=1):
+    cpdef void set_param_value(self, int index, float value, float x_min=0, float x_max=1):
         """Set the value of the parameter at a give index and check beforehand if it is valid
         
         Args:
             index (int): Index of the parameter
-            value (double): Value that is changed
-            x_min (double): Min value of the domain for the x axis
-            x_max (double): Max value of the domain for the x axis
+            value (float): Value that is changed
+            x_min (float): Min value of the domain for the x axis
+            x_max (float): Max value of the domain for the x axis
 
         Raises:
             Exception: The index is out of bounds or the value is invalid for the corresponding parameter
@@ -148,23 +148,23 @@ cdef class AbstractMF:
             return False
         return np.array_equal(self._params, other.get_params())
 
-    cpdef cnp.ndarray[double, ndim=2] get_plot_points(self, double x_min=0, double x_max=1):
+    cpdef cnp.ndarray[float, ndim=2] get_plot_points(self, float x_min=0, float x_max=1):
         """Get the plot points coordinates
         
         Args:
-            x_min (double): Min value of the domain for the x axis
-            x_max (double): Max value of the domain for the x axis
+            x_min (float): Min value of the domain for the x axis
+            x_max (float): Max value of the domain for the x axis
         Returns:
             Points coordinates that define this function shape
         """
         raise AbstractMethodException()
 
-    cpdef double get_support(self, double x_min=0, double x_max=0):
+    cpdef float get_support(self, float x_min=0, float x_max=0):
         """Get the support value associated to this function: area covered by this function in the space "domain x [0, 1]"
         
         Args:
-            x_min (double): Min value of the domain for the x axis
-            x_max (double): Max value of the domain for the x axis
+            x_min (float): Min value of the domain for the x axis
+            x_max (float): Max value of the domain for the x axis
             
         Returns:
             Support value
