+++ src/mofgbmlpy/fuzzy/fuzzy_term/membership_function/triangular_mf.pxd
@@ -3,8 +3,8 @@ import cython
 cimport numpy as cnp
 
 cdef class TriangularMF(AbstractMF):
-    cdef double get_value(self, double x)
-    cpdef cnp.ndarray[double, ndim=1] get_param_range(self, int index, double x_min=?, double x_max=?)
-    cpdef bint is_param_value_valid(self, int index, double value, double x_min=?, double x_max=?)
-    cpdef cnp.ndarray[double, ndim=2] get_plot_points(self, double x_min=?, double x_max=?)
-    cpdef double get_support(self, double x_min=?, double x_max=?)
+    cdef float get_value(self, float x)
+    cpdef cnp.ndarray[float, ndim=1] get_param_range(self, int index, float x_min=?, float x_max=?)
+    cpdef bint is_param_value_valid(self, int index, float value, float x_min=?, float x_max=?)
+    cpdef cnp.ndarray[float, ndim=2] get_plot_points(self, float x_min=?, float x_max=?)
+    cpdef float get_support(self, float x_min=?, float x_max=?)
