+++ src/mofgbmlpy/fuzzy/fuzzy_term/membership_function/triangular_mf.pyx
@@ -12,9 +12,9 @@ cdef class TriangularMF(AbstractMF):
         """Constructor
 
         Args:
-            left (double): X coordinate of the leftmost vertex of the triangle: membership is equals to 0 before it
-            center (double): X coordinate of the vertex in the center of the triangle: membership is equals to 1 at this point
-            right (double): X coordinate of the leftmost vertex of the triangle: membership is equals to 0 after it
+            left (float): X coordinate of the leftmost vertex of the triangle: membership is equals to 0 before it
+            center (float): X coordinate of the vertex in the center of the triangle: membership is equals to 1 at this point
+            right (float): X coordinate of the leftmost vertex of the triangle: membership is equals to 0 after it
         """
         if left is None or center is None or right is None:
             raise TypeError("Parameters can't be None")
@@ -23,16 +23,16 @@ cdef class TriangularMF(AbstractMF):
         elif center > right:
             raise ValueError(f"Error in triangular membership function: center={center:.2f} should be <= right={right:.2f}")
 
-        super().__init__(np.array([left,center,right], dtype=np.float64))
+        super().__init__(np.array([left,center,right], dtype=np.float32))
 
-    cdef double get_value(self, double x):
+    cdef float get_value(self, float x):
         """Get membership value (accessible only from Cython code)
 
         Args:
-            x (double): Value whose membership value is calculated
+            x (float): Value whose membership value is calculated
 
         Returns:
-            double: Membership value
+            float: Membership value
         """
         if x == self._params[1]:
             # For the case where left = center or center = right
@@ -53,26 +53,26 @@ cdef class TriangularMF(AbstractMF):
         """
         return "<Triangular MF (%f, %f, %f)>" % (self._params[0], self._params[1], self._params[2])
 
-    cpdef cnp.ndarray[double, ndim=1] get_param_range(self, int index, double x_min=0, double x_max=1):
+    cpdef cnp.ndarray[float, ndim=1] get_param_range(self, int index, float x_min=0, float x_max=1):
         """Get the range of acceptable values a given parameter as a numpy array of two values
 
             Args:
                 index (int): Index of the parameter whose range is got
-                x_min (double): Min value of the domain for the x axis (e.g. if index is 0 for the triangular set then we get [xmin, center]
-                x_max (double): Max value of the domain for the x axis (e.g. if index is 2 for the triangular set then we get [center, max]
+                x_min (float): Min value of the domain for the x axis (e.g. if index is 0 for the triangular set then we get [xmin, center]
+                x_max (float): Max value of the domain for the x axis (e.g. if index is 2 for the triangular set then we get [center, max]
 
             Returns:
-                double[]: Range of possible values
+                float[]: Range of possible values
             """
         if x_min > self._params[0] or x_max < self._params[2]:
             raise ValueError(f"Invalid x_min or x_max. They must be in the range [{self._params[0]}, {self._params[2]}]")
 
         if index == 0:
-            return np.array([x_min, self._params[1]])
+            return np.array([x_min, self._params[1]], dtype=np.float32)
         elif index == 1:
-            return np.array([self._params[0], self._params[2]])
+            return np.array([self._params[0], self._params[2]], dtype=np.float32)
         elif index == 2:
-            return np.array([self._params[1], x_max])
+            return np.array([self._params[1], x_max], dtype=np.float32)
         else:
             raise IndexError("Invalid index for rectangular MF")
 
@@ -90,12 +90,12 @@ cdef class TriangularMF(AbstractMF):
         memo[id(self)] = new_object
         return new_object
 
-    cpdef cnp.ndarray[double, ndim=2] get_plot_points(self, double x_min=0, double x_max=1):
+    cpdef cnp.ndarray[float, ndim=2] get_plot_points(self, float x_min=0, float x_max=1):
         """Get the plot points coordinates
 
            Args:
-               x_min (double): Min value of the domain for the x axis
-               x_max (double): Max value of the domain for the x axis
+               x_min (float): Min value of the domain for the x axis
+               x_max (float): Max value of the domain for the x axis
            Returns:
                Points coordinates that define this function shape
            """
@@ -106,14 +106,14 @@ cdef class TriangularMF(AbstractMF):
             [self._params[1], 1],
             [self._params[2], 0],
             [x_max, 0],
-        ], np.float64)
+        ], np.float32)
 
-    cpdef double get_support(self, double x_min=0, double x_max=0):
+    cpdef float get_support(self, float x_min=0, float x_max=0):
         """Get the support value associated to this function: area covered by this function in the space "domain x [0, 1]"
 
         Args:
-            x_min (double): Min value of the domain for the x axis
-            x_max (double): Max value of the domain for the x axis
+            x_min (float): Min value of the domain for the x axis
+            x_max (float): Max value of the domain for the x axis
 
         Returns:
             Support value
